Move covariance builder under sim package
