+++ pa_core/sim/__init__.py
@@ -7,10 +7,12 @@ from .paths import (
     draw_financing_series,
     simulate_alpha_streams,
 )
+from .covariance import build_cov_matrix
 
 __all__ = [
     "simulate_financing",
     "prepare_mc_universe",
+    "build_cov_matrix",
     "draw_joint_returns",
     "draw_financing_series",
     "simulate_alpha_streams",
